requirements.txt
faiss-cpu
groq
langchain-groq
pyPDF2
langchain_google_genai
langchain
streamlit
langchain_community
python-dotenv
pypdf
google-cloud-aiplatform>=1.38
